mvp2 - double -> BigDecimal for amount

+++ app/src/main/java/spending/tracker/android/domain/repository/SpendingRepository.kt
@@ -2,6 +2,7 @@ package spending.tracker.android.domain.repository
 
 import kotlinx.coroutines.flow.Flow
 import spending.tracker.android.domain.model.Spending
+import java.math.BigDecimal
 
 /**
  * Offline-first репозиторий расходов.
@@ -28,7 +29,7 @@ interface SpendingRepository {
      */
     suspend fun addSpending(
         userEmail: String,
-        amount: Double,
+        amount: BigDecimal,
         categoryId: Long,
         subCategoryId: Long?,
         description: String?,
@@ -38,7 +39,7 @@ interface SpendingRepository {
     suspend fun updateSpending(
         userEmail: String,
         id: Long,
-        amount: Double,
+        amount: BigDecimal,
         categoryId: Long,
         subCategoryId: Long?,
         description: String?,
